A Bimodal Weibull Distribution

Density, distribution function, quantile function and random generation for a Bimodal Weibull Distribution with parameters shape and scale.

Usage

dbwd(x, alpha, beta = 1, sigma, log = FALSE)

pbwd(q, alpha, beta = 1, sigma, lower.tail = TRUE, log.p = FALSE)

qbwd(p, alpha, beta = 1, sigma, lower.tail = TRUE)

rbwd(n, alpha, beta = 1, sigma)

Arguments

x, q

vector of quantiles.

alpha

a shape parameter.

beta

a scale parameter.

sigma

the parameter that controls the uni- or bimodality of the distribution.

log, log.p

logical; if TRUE, probabilities p are given as log(p).

lower.tail

logical; if TRUE (default), probabilities are \(P\left[ X\leq x\right]\), otherwise,\(P\left[ X>x\right] \).

p

vector of probabilities.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

Value

dbwd gives the density, pbwd gives the distribution function, qbwd gives the quantile function and rbwd generates random deviates.

Details

A Bimodal Weibull distribution with shape parameter \(\alpha\), scale parameter \(\beta\) and the parameter that controls the uni- or bimodality of the distribution \(\sigma\), has density given by $$f\left( x\right) =\frac{\alpha }{\beta Z_{\theta }} \left[ 1+\left( 1-\sigma~x\right) ^{2}\right] \left( \frac{x}{\beta } \right) ^{\alpha -1}\exp \left( -\left( \frac{x}{\beta }\right) ^{\alpha } \right),$$ where $$Z_{\theta }=2+\sigma ^{2}\beta ^{2}\Gamma \left( 1+\left( 2/\alpha \right)\right) -2\sigma \beta \Gamma \left( 1+\left( 1/\alpha \right) \right) $$ and $$x\geq 0,~\alpha ,\beta >0,~\sigma \in\mathbb{R}.$$

References

Vila, R. ve Niyazi Çankaya, M., 2022, A bimodal Weibull distribution: properties and inference, Journal of Applied Statistics, 49 (12), 3044-3062.

Examples

library(new.dist)
dbwd(1,alpha=2,beta=3,sigma=4)
#> [1] 0.01594262
pbwd(1,alpha=2,beta=3,sigma=4)
#> [1] 0.003859685
qbwd(.7,alpha=2,beta=3,sigma=4)
#> [1] 4.759942
rbwd(10,alpha=2,beta=3,sigma=4)
#>  [1] 2.675128 5.389236 1.716144 4.518051 4.321859 4.434323 8.718917 4.618163
#>  [9] 4.005316 6.129080