The Power Muth Distribution

Density, distribution function, quantile function and random generation for the Power Muth distribution with parameters shape and scale.

Usage

dtpmd(x, beta = 1, alpha, log = FALSE)

ptpmd(x, beta = 1, alpha, lower.tail = TRUE, log.p = FALSE)

qtpmd(p, beta = 1, alpha, lower.tail = TRUE)

rtpmd(n, beta = 1, alpha)

Arguments

x

vector of quantiles.

beta

a scale parameter.

alpha

a shape parameter.

log, log.p

logical; if TRUE, probabilities p are given as log(p).

lower.tail

logical; if TRUE (default), probabilities are \(P\left[ X\leq x\right]\), otherwise,\(P\left[ X>x\right] \).

p

vector of probabilities.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

Value

dtpmd gives the density, ptpmd gives the distribution function, qtpmd gives the quantile function and rtpmd generates random deviates.

Details

The Power Muth Distribution with shape parameter \(\alpha\) and scale parameter \(\beta\) has density given by $$f\left( x\right) =\frac{\alpha }{\beta ^\alpha }x^{\alpha -1} \left( e^{\left(x/\beta \right) ^{\alpha }}-1\right) \left( e^{\left( x/\beta \right) ^{\alpha }- \left( e^{\left( x/\beta \right) ^{\alpha }}-1\right) }\right), $$ where $$x>0,~\alpha ,\beta>0.$$

Note

Hazard function; $$h\left( \beta ,\alpha \right) =\frac{\alpha }{\beta ^{\alpha }} \left(e^{\left( x/\beta \right) ^{\alpha }}-1\right) x^{\alpha -1}$$

References

Jodra, P., Gomez, H. W., Jimenez-Gamero, M. D., & Alba-Fernandez, M. V. (2017). The power Muth distribution . Mathematical Modelling and Analysis, 22(2), 186-201.

Examples

library(new.dist)
dtpmd(1, beta=2, alpha=3)
#> [1] 0.04952547
ptpmd(1,beta=2,alpha=3)
#> [1] 0.008115344
qtpmd(.5,beta=2,alpha=3)
#> [1] 1.990084
rtpmd(10,beta=2,alpha=3)
#>  [1] 1.651854 1.889539 1.978503 1.411612 2.025935 2.407632 1.922025 1.379525
#>  [9] 1.751835 1.875832